added atom mapping to lammps
